Support 64 bit field for fstat
Summary:
Some of the field of "struct stat" are eight byte longs.
Thus avoid loosing information by casting 64 bits data to 32 bits.

Test Plan: Ensure that fstat is working properlly in openamp

Reviewers: O26 newlib, jhascoet

Reviewed By: O26 newlib, jhascoet

Differential Revision: https://phab.kalray.eu/D2318

+++ libgloss/k1-cos/fstat.c
@@ -11,7 +11,7 @@ int sc_fstat(int fd, int *res);
 static inline int __gloss_fstat(int fd, struct stat *st)
 {
   int ret;
-  int res[13];
+  long res[13];
 
   ret = sc_fstat(fd, res);
